fixed links rendering bug
this fixes the links rendering bug by simply putting the filled nodes
on top of the links.
To properly fix this, we´d need to address the fact that sometimes the
link is thicker than the available width between 2 columns. One way
would be to have a separate rendering mode for thick links + high ys
delta, drawing a filled path thinner in the middle, instead of a simple
stroke.