fix home page dropdowns on Firefox
@andresgnlez @hectoruch querySelectorAll does not return an Array but a
NodeList. You cannot use forEach and other array methods on NodeLists.
Unfortunately Chrome supports this even if it’s not in the spec so we
all get fucked, but be careful with this because it doesn’t work on FF
and Safari.
The fix is converting the NodeList to an array using
Array.prototype.slice.call(nodeList, 0)